requirements.txt
streamlit
pyserial
selenium
webdriver-manager